Move js files to js module directory
