+++ src/Jsmodules/displayPokemon.js
@@ -1,4 +1,4 @@
-import addeventComment from '../display-comment.js';
+import addeventComment from './display-comment.js';
 import { getLikes, addeventLikes } from './displaylikes.js';
 
 const pokemonContainer = document.querySelector('.pokemon-container');
